Allow Executor::run_and_wait() and loop_until() from non-worker thread
Not allowing run_and_wait() and loop_until() to be called from
non-worker threads is problematic, because the only other way to wait
for completion of a Taskflow instance involve actual blocking on
std::future outside of taskflow control. If this happens in nested
context, then the thread being blocked is a worker thread. If this
happens enough times, all threads managed by taskflow will be blocked
which will result in a deadlock.

To prevent this, the user must use `run_and_wait()` within threads
managed by taskflow and regular `run()` outside. It is unfortunately
often not possible to know whether thread is managed by taskflow,
especially in larger projects that consists of many sub-components that
have their own nested parallelization.

A solution to this is to remove the limitation for running
Executor::run_and_wait() and loop_until() from non-worker thread. The
implementation simply creates a Taskflow instance, sends tasks to be
performed by taskflow threads and blocks on the returned std::future.
This is exactly how the same situation is usually implemented in user
code.